\name{dist.pvclust4PAN}
\alias{dist.pvclust4PAN}
\title{
Compute distances between gene phenotypes for pvclust
}
\description{
This is a function for PAN to compute distances between phenotypes of two genes. 
This function modifies the function dist.pvclust from R package pvclust by 
adding in first- and second-order cosine distances.
}
\usage{
##see package \code{\link[pvclust:pvclust]{pvclust}} for details
}

\author{
Xin Wang
}

