\name{parPvclust4PAN}
\alias{parPvclust4PAN}
\title{
Compute distances between gene phenotypes for pvclust
}
\description{
This is a function for PAN to do parallel computing for \code{pvclust}.
This function modifies the function parPvclust from R package pvclust to allow 
clustering on first- and second-order cosine distances.
}
\usage{
##see \code{\link[pvclust:parPvclust]{parPvclust}} for details
}
\author{
Xin Wang
}

